# improve perf of Enum.Humanize