--- a/src/Humanizer.Tests.Shared/EnumHumanizeTests.cs
+++ b/src/Humanizer.Tests.Shared/EnumHumanizeTests.cs
@@ -49,4 +49,108 @@ public class EnumHumanizeTests
     [Fact]
     public void HonorsLocalizedDisplayAttribute() =>
         Assert.Equal(EnumTestsResources.MemberWithLocalizedDisplayAttribute, EnumUnderTest.MemberWithLocalizedDisplayAttribute.Humanize());
+
+    [Fact]
+    public void HumanizeCustomPropertyAttributeWithLocator()
+    {
+        Configurator.EnumDescriptionPropertyLocator = p => p.Name == "Info";
+        try
+        {
+            Assert.Equal(EnumTestsResources.MemberWithCustomPropertyAttribute, EnumForCustomLocator.MemberWithCustomPropertyAttribute.Humanize());
+        }
+        finally
+        {
+            Configurator.EnumDescriptionPropertyLocator = null;
+        }
+    }
+
+    [Fact]
+    public void HumanizeMembersWithoutDescriptionAttributeWithLocator()
+    {
+        Configurator.EnumDescriptionPropertyLocator = p => p.Name == "Info";
+        try
+        {
+            Assert.Equal(EnumTestsResources.MemberWithoutDescriptionAttributeSentence, EnumForCustomLocator.MemberWithoutDescriptionAttribute.Humanize());
+        }
+        finally
+        {
+            Configurator.EnumDescriptionPropertyLocator = null;
+        }
+    }
+    [Fact]
+    public void DehumanizeThrowsForEnumNoMatch()
+    {
+        Assert.Throws<NoMatchFoundException>(() => EnumTestsResources.MemberWithDescriptionAttribute.DehumanizeTo<DummyEnum>());
+        Assert.Throws<NoMatchFoundException>(() => EnumTestsResources.MemberWithDescriptionAttribute.DehumanizeTo(typeof(DummyEnum)));
+    }
+
+    [Fact]
+    public void DehumanizeCanReturnNullForEnumNoMatch() =>
+        Assert.Null(EnumTestsResources.MemberWithDescriptionAttribute.DehumanizeTo(typeof(DummyEnum), OnNoMatch.ReturnsNull));
+
+    [Fact]
+    public void DehumanizeDescriptionAttribute()
+    {
+        Assert.Equal(EnumUnderTest.MemberWithDescriptionAttribute, EnumTestsResources.MemberWithDescriptionAttribute.DehumanizeTo<EnumUnderTest>());
+        Assert.Equal(EnumUnderTest.MemberWithDescriptionAttribute, EnumTestsResources.MemberWithDescriptionAttribute.DehumanizeTo(typeof(EnumUnderTest)));
+    }
+
+    [Fact]
+    public void DehumanizeDescriptionAttributeSubclasses()
+    {
+        const string calculatedDescription = "Overridden " + EnumTestsResources.MemberWithDescriptionAttributeSubclass;
+        Assert.Equal(EnumUnderTest.MemberWithDescriptionAttributeSubclass, calculatedDescription.DehumanizeTo<EnumUnderTest>());
+        Assert.Equal(EnumUnderTest.MemberWithDescriptionAttributeSubclass, calculatedDescription.DehumanizeTo(typeof(EnumUnderTest)));
+    }
+
+    [Fact]
+    public void DehumanizeAnyAttributeWithDescriptionStringProperty()
+    {
+        Assert.Equal(EnumUnderTest.MemberWithCustomDescriptionAttribute, EnumTestsResources.MemberWithCustomDescriptionAttribute.DehumanizeTo<EnumUnderTest>());
+        Assert.Equal(EnumUnderTest.MemberWithCustomDescriptionAttribute, EnumTestsResources.MemberWithCustomDescriptionAttribute.DehumanizeTo(typeof(EnumUnderTest)));
+    }
+
+    [Fact]
+    public void DehumanizeMembersWithoutDescriptionAttribute()
+    {
+        Assert.Equal(EnumUnderTest.MemberWithoutDescriptionAttribute, EnumTestsResources.MemberWithoutDescriptionAttributeSentence.DehumanizeTo<EnumUnderTest>());
+        Assert.Equal(EnumUnderTest.MemberWithoutDescriptionAttribute, EnumTestsResources.MemberWithoutDescriptionAttributeSentence.DehumanizeTo(typeof(EnumUnderTest)));
+    }
+
+    [Theory]
+    [InlineData(EnumTestsResources.MemberWithoutDescriptionAttributeTitle, EnumUnderTest.MemberWithoutDescriptionAttribute)]
+    [InlineData(EnumTestsResources.MemberWithoutDescriptionAttributeLowerCase, EnumUnderTest.MemberWithoutDescriptionAttribute)]
+    [InlineData(EnumTestsResources.MemberWithoutDescriptionAttributeSentence, EnumUnderTest.MemberWithoutDescriptionAttribute)]
+    public void DehumanizeIsCaseInsensitive(string input, EnumUnderTest expectedEnum)
+    {
+        Assert.Equal(expectedEnum, input.DehumanizeTo<EnumUnderTest>());
+        Assert.Equal(expectedEnum, input.DehumanizeTo(typeof(EnumUnderTest)));
+    }
+
+    [Fact]
+    public void DehumanizeAllCapitalMembersAreReturnedAsIs()
+    {
+        Assert.Equal(EnumUnderTest.ALLCAPITALS, EnumUnderTest.ALLCAPITALS.ToString().DehumanizeTo<EnumUnderTest>());
+        Assert.Equal(EnumUnderTest.ALLCAPITALS, EnumUnderTest.ALLCAPITALS.ToString().DehumanizeTo(typeof(EnumUnderTest)));
+    }
+
+    [Fact]
+    public void DehumanizeDisplayAttribute()
+    {
+        Assert.Equal(EnumUnderTest.MemberWithDisplayAttribute, EnumTestsResources.MemberWithDisplayAttribute.DehumanizeTo<EnumUnderTest>());
+        Assert.Equal(EnumUnderTest.MemberWithDisplayAttribute, EnumTestsResources.MemberWithDisplayAttribute.DehumanizeTo(typeof(EnumUnderTest)));
+    }
+
+    [Fact]
+    public void DehumanizeLocalizedDisplayAttribute()
+    {
+        Assert.Equal(EnumUnderTest.MemberWithLocalizedDisplayAttribute, EnumTestsResources.MemberWithLocalizedDisplayAttribute.DehumanizeTo<EnumUnderTest>());
+        Assert.Equal(EnumUnderTest.MemberWithLocalizedDisplayAttribute, EnumTestsResources.MemberWithLocalizedDisplayAttribute.DehumanizeTo(typeof(EnumUnderTest)));
+    }
+
+    enum DummyEnum
+    {
+        First,
+        Second
+    }
 }