… src/Benchmarks/EnumBenchmarks.cs
… src/Humanizer.Tests/ApiApprover/PublicApiApprovalTest.Approve_Public_Api.DotNet8_0.verified.txt
… src/Humanizer.Tests/ApiApprover/PublicApiApprovalTest.Approve_Public_Api.Net4_8.verified.txt
… src/Humanizer/EnumDehumanizeExtensions.cs