## Enum.Humanize

### Version 2.14.1

| Method        | Mean     | Error     | StdDev    | Allocated |
|-------------- |---------:|----------:|----------:|----------:|
| Enum.Humanize | 5.363 us | 0.0182 us | 0.0152 us |     848 B |

### Version 3

| Method        | Mean     | Error    | StdDev   | Allocated |
|-------------- |---------:|---------:|---------:|----------:|
| Enum.Humanize | 32.08 ns | 0.081 ns | 0.072 ns |      24 B |

## Enum.DehumanizeTo 

### Version 2.14.1

| Method            | Mean     | Error    | StdDev   | Allocated |
|------------------ |---------:|---------:|---------:|----------:|
| Enum.DehumanizeTo | 21.42 us | 0.250 us | 0.234 us |   4.72 KB |

### Version 3

| Method            | Mean     | Error    | StdDev   | Allocated |
|------------------ |---------:|---------:|---------:|----------:|
| Enum.DehumanizeTo | 62.26 ns | 0.344 ns | 0.305 ns |         - |